LIBASPACE-210. Removed "flash" request instructions
Removed use of Rails "flash" functionality for displaying the request
instructions on collection pages.

Since the intention is that the request instructions appear consistently
on all collection pages, moved the request instructions into the
appropriate ERB templates in the "umd-lib-aspace-theme" plugin.

Modified the "container" page to use the "request_instructions" CSS
rule (specified in the "umd-lib-aspace-theme" plugin) for styling.

Modified the request instructions text for collections and containers
as specified in the Google document (see LIBASPACE-211).

https://issues.umd.edu/browse/LIBASPACE-210